Introduce offline mode
If system administrators need to they can set the whole Turnitin plugin
to offline mode. This can be desirable when networking operations
are taking up too long to complete, blocking valuable Apache processes
while doing so. The system administrator can temporary disable Turnitin
to free up resources while the networking problem is diagnosed.

When offline mode is enabled users will be presented with an error on
turnitin pages:

    "We are experiencing a temporary problem. Please try again shortly"

Offline mode is enabled adding the following to config.php for Moodle:

    $CFG->tiioffline = true;

The Test Turnitin Connection action on the settings page is unaffected
by the offline mode. So that action can still be used by system
administrators when troubleshooting any connection problems.