Setting dates on a part should work when Location setting is not set to 'Server's local time'
On a CentOS server with its timezone set to "Europe/London" during daylight savings the timezone is reported as BST, while during non-daylightsavings it is reported as GMT. This causes issues with dates set for daylight savings times while the server is in a non-daylight savings time. The bugs caused by this can be worked around by setting Moodle's Location settings to have a timezone of "Europe/London"

Unfortunately in this configuration it seems that setting part dates via the turnitintooltwo AJAX causes times during daylight savings to be incorrect.

This change does two things:

* Ensures that the date string passed via AJAX is changed into the correct timestamp based on the timezone settings of the user.
* Removes a call to userdate as the timestamp passed should always be UCT already at this point.