avoid endless loops creating temporary files
Firstly, I added a safety valve to prevent endless loops from being
a concern if something bizarre prevents files being created. After 10
attempts with different random values it throws an exception.

Secondly, previously, filenames having no sensible extension could
cause an endless loop as the filename was deemed the extension, and
the concatenation of it to the constructed name produces something
the OS rejects. This changes the logic to at least stay within the
limits, preserving the extension as much as possible. Someone can
still do silly things like name a file "Part 1.1 name <snip 100ch>
here" making the extension appear to be "1 name <snip 100ch> here",
but it will not cause the file path length to be erroneous.