Fix sort order of Submission Part Tab
The submission parts can don't always apply in the same order they have
been created because there isn't a part number to sort them by, and
without a order by clause, the DB is not guaranteed to return the
records in any specific order.

Previously in Turnitintool v1, the submissions tabs were ordered by
first to last created, so this change mimics that.